add type_map.rs
