+++ src/rest/mod.rs
@@ -1,4 +1,4 @@
 mod endpoint;
 mod endpoint_set;
-
+mod type_map;
 pub use endpoint_set::EndpointSet;
